Switch budget/period filters to use BehaviourSubject
Instead of ReplaySubject, as this is causing troubles while initialising
on mobile devices, where old subscriptions are kicking in and filters
gets confused.

With BehaviorSubject, when session service emits, that session is deactivated,
which happens on first run of the app, budgets are emptied, which triggers budgets
subscription in timeframe service, but this one can't proceed, as current budget did
not emit yet.

Alternative to this would be to cancel all current subscriptions after session is activated,
but just switching to BehaviorSubject is simpler and sufficient, as subscriptions can finish,
so we get again clear state once user logs in.